Consolidate shared ports module
