+++ apps/server/vibesensor/shared/__init__.py
@@ -1,1 +1,21 @@
 """Cross-cutting shared server helpers."""
+
+from vibesensor.shared.ports import (
+    ClientTracker,
+    ResolvedSpeedSnapshot,
+    RunPersistence,
+    SettingsReader,
+    SignalSource,
+    SpeedProvider,
+    TrackedClient,
+)
+
+__all__ = [
+    "ClientTracker",
+    "ResolvedSpeedSnapshot",
+    "RunPersistence",
+    "SettingsReader",
+    "SignalSource",
+    "SpeedProvider",
+    "TrackedClient",
+]
