+++ apps/server/vibesensor/shared/types/__init__.py
@@ -1,1 +1,1 @@
-"""Shared typed payloads, ports, and API models."""
+"""Shared typed payloads and API models."""
