+++ apps/server/vibesensor/use_cases/history/runs.py
@@ -8,10 +8,9 @@ from typing import Never
 from vibesensor.domain import RunStatus
 from vibesensor.shared.boundaries.diagnostic_case import project_analysis_summary
 from vibesensor.shared.exceptions import AnalysisNotReadyError, RunNotFoundError
+from vibesensor.shared.ports import RunPersistence, SettingsReader
 from vibesensor.shared.run_context import add_current_context_warnings, localize_warning_list
 from vibesensor.shared.types.json_types import JsonObject, is_json_object
-from vibesensor.shared.types.run_persistence import RunPersistence
-from vibesensor.shared.types.settings_reader import SettingsReader
 from vibesensor.use_cases.history.helpers import (
     HistoryRecord,
     async_require_run,
