+++ apps/server/vibesensor/use_cases/run/logger.py
@@ -26,14 +26,16 @@ from uuid import uuid4
 
 from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
 from vibesensor.shared.constants import NUMERIC_TYPES
+from vibesensor.shared.ports import (
+    ClientTracker,
+    RunPersistence,
+    SettingsReader,
+    SignalSource,
+    SpeedProvider,
+)
 from vibesensor.shared.time_utils import utc_now_iso
-from vibesensor.shared.types.client_tracker import ClientTracker
 from vibesensor.shared.types.health_snapshot import RunRecorderHealthSnapshot
 from vibesensor.shared.types.json_types import JsonObject
-from vibesensor.shared.types.run_persistence import RunPersistence
-from vibesensor.shared.types.settings_reader import SettingsReader
-from vibesensor.shared.types.signal_source import SignalSource
-from vibesensor.shared.types.speed_provider import SpeedProvider
 from vibesensor.use_cases.run.lifecycle_state import ActiveRunSnapshot, RunLifecycleState
 from vibesensor.use_cases.run.persistence_writer import (
     _APPEND_RETRY_DELAYS_S,
